app/Llm_cpu/requirements.txt
transformers
gguf>=0.1.0
wget==3.2
torch==2.0.1

tokenizers==0.13.3
Flask==2.3.3
wget==3.2
wheel==0.38.4
make==0.1.6.post2
huggingface-hub==0.17.1
Jinja2==3.1.2
numpy==1.25.2
safetensors==0.3.3

sentencepiece==0.1.98
requests==2.29.0

networkx==3.1
sympy==1.12
tqdm==4.65.0
arrow==1.2.3
colorama==0.4.6
fsspec==2023.9.0
PyYAML==6.0.1
regex==2023.8.8
accelerate==0.23.0
arrow==1.2.3
boltons==23.0.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
cmake==3.27.4.1
colorama==0.4.6
accelerate==0.23.0
arrow==1.2.3
brotlipy==0.7.0
filelock==3.12.4
fsspec==2023.9.0
jinja2-time==0.2.0
jsonpointer==2.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
psutil==5.9.5
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
sympy==1.12
typing_extensions==4.7.1
yarg==0.1.9
zstandard==0.19.0


#conda==23.5.2
#transformers==4.34.0.dev0